Updated atbanalyticsgrp_dark.py,utils.py, homepage.py, fundamentalAnalysis.py, technicalAnalysis.py
Created and updated the color scales to use variable based objects instead of hard-coding
Changed Heatmap to reflect aqua_to_rosy coloring instead of gold_to_aqua
Added get_random_color() to utils.py for generating random branding colors
Updated all_price_sql to stocks_in_sector_sql and have it reading through dedicated SQL file
Updated Security Returns to be all the stocks in the sector that is chosen
Added the Company Dropdown and Stock input to the fundamentalAnalysis.py and technicalAnalysis.py
Added company_info_df.query to company_info_df when we have a stock symbol in fundamentalAnalysis.py so that the proper stock is selected and not just the first one
Changed the layout of the homepage to include more visuals and less text
Added updated images from unsplash to the assets folder
Created QuickRatios.sql
Created Financial Definitions.xlsx to be used in helping understand some of the ratios present
updated max_date to include the ms(.%f) to stop the ValueError from happening.
Corrected defaults in MyStock init function
Added Stock Selection UI to technicalAnalysis.py
Updated callback by commenting out the stock info(name,title,price,etc) in the Output as well as the Update_UI return statement.

I know this is a wayyyyyyyyy to long of a commit but this is the first commit after months away.For Shameee Sir, For Shameeeee.